Refactored code to split off PagedVolume from PagedVolumeComponent.
This is setting up for components such as 'LandWaterVolume', which
(as its name implies) contains a Volume containing both land and water.
This would require two separate volumes -- one for the land, and one for
the water, which 'flows' downhill based on the what the shape of the land.

This functionality is not in place yet, but should be later.